Common, UT: Move get_file_sha256_hash  to unit tests
